refactored existing parser into `scala_parser_combinators` package
