+++ parser/src/main/scala/eo/parser/scala_parser_combinators/Tokens.scala
@@ -1,4 +1,4 @@
-package eo.parser
+package eo.parser.scala_parser_combinators
 
 sealed trait Token
