+++ utils/src/main/scala/eo/utils/text.scala
@@ -1,9 +1,14 @@
 package eo.utils
 
 object text {
-  def indentWith(c: Char)(n: Int)(s: String) = s"${c.toString * n}${s}"
+  def indentWith(c: Char)(n: Int)(s: String) = s"${c.toString * n}$s"
 
   def indentSpace: Int => String => String = indentWith(' ')
 
   def indent: String => String = indentWith(' ')(2)
+
+  def main(args: Array[String]): Unit = {
+    println(indent("hello"))
+  }
+
 }
